Using lseek64

+++ main.cpp
@@ -249,10 +249,10 @@ int main(int argc, char *argv[])
                 diff.compare();
             } else if (address) {
                 cout << "Cluster " << cluster << " address:" << endl;
-                int addr = fat.clusterAddress(cluster);
+                long long addr = fat.clusterAddress(cluster);
                 int next1 = fat.nextCluster(cluster, 0);
                 int next2 = fat.nextCluster(cluster, 1);
-                printf("%d (%08x)\n", addr, addr);
+                printf("%llu (%08x)\n", addr, addr);
                 cout << "Next cluster:" << endl;
                 printf("FAT1: %u (%08x)\n", next1, next1);
                 printf("FAT2: %u (%08x)\n", next2, next2);
